+++ src/allotropy/allotrope/schema_mappers/adm/plate_reader/benchling/_2023/_09/plate_reader.py
@@ -128,6 +128,7 @@ class Measurement:
     well_plate_identifier: str | None = None
     detection_type: str | None = None
     sample_role_type: SampleRoleType | None = None
+    firmware_version: str | None = None
 
     # Measurements
     absorbance: JsonFloat | None = None
@@ -280,6 +281,7 @@ class Mapper(SchemaMapper[Data, Model]):
     ) -> OpticalImagingMeasurementDocumentItems:
         device_control_document = OpticalImagingDeviceControlDocumentItem(
             device_type=measurement.device_type,
+            firmware_version=measurement.firmware_version,
             detection_type=measurement.detection_type,
             detector_wavelength_setting=quantity_or_none(
                 TQuantityValueNanometer,
@@ -315,6 +317,10 @@ class Mapper(SchemaMapper[Data, Model]):
                 measurement.excitation_wavelength_setting,
             ),
         )
+        # TODO(ASM gaps): we think this should be added to ASM
+        custom_info_doc = {
+            "LED filter": measurement.led_filter,
+        }
 
         return OpticalImagingMeasurementDocumentItems(
             measurement_identifier=measurement.identifier,
@@ -322,13 +328,8 @@ class Mapper(SchemaMapper[Data, Model]):
             device_control_aggregate_document=OpticalImagingDeviceControlAggregateDocument(
                 device_control_document=[
                     add_custom_information_document(
-                        device_control_document,
-                        {
-                            "LED filter": measurement.led_filter,
-                        },
+                        device_control_document, custom_info_doc
                     )
-                    if measurement.led_filter
-                    else device_control_document
                 ]
             ),
             processed_data_aggregate_document=self._get_processed_data_aggregate_document(
@@ -346,6 +347,7 @@ class Mapper(SchemaMapper[Data, Model]):
                 device_control_document=[
                     UltravioletAbsorbancePointDetectionDeviceControlDocumentItem(
                         device_type=measurement.device_type,
+                        firmware_version=measurement.firmware_version,
                         detection_type=measurement.detection_type,
                         detector_wavelength_setting=quantity_or_none(
                             TQuantityValueNanometer,
@@ -384,6 +386,7 @@ class Mapper(SchemaMapper[Data, Model]):
                 device_control_document=[
                     LuminescencePointDetectionDeviceControlDocumentItem(
                         device_type=measurement.device_type,
+                        firmware_version=measurement.firmware_version,
                         detection_type=measurement.detection_type,
                         detector_wavelength_setting=quantity_or_none(
                             TQuantityValueNanometer,
@@ -423,6 +426,7 @@ class Mapper(SchemaMapper[Data, Model]):
                 device_control_document=[
                     FluorescencePointDetectionDeviceControlDocumentItem(
                         device_type=measurement.device_type,
+                        firmware_version=measurement.firmware_version,
                         detection_type=measurement.detection_type,
                         detector_wavelength_setting=quantity_or_none(
                             TQuantityValueNanometer,