+++ src/allotropy/parsers/appbio_quantstudio/appbio_quantstudio_structure.py
@@ -118,11 +118,8 @@ class WellItem(Referenceable):
     well_location_identifier: str | None = None
     quencher_dye_setting: str | None = None
     sample_role_type: str | None = None
-    sample_color: str | None = None
-    biogroup_name: str | None = None
-    biogroup_color: str | None = None
-    target_name: str | None = None
-    target_color: str | None = None
+    group_identifier: str | None = None
+    extra_data: dict[str, Any] | None = None
     _result: Result | None = None
 
     # Make hashable to allow for use of caching
@@ -155,11 +152,14 @@ class WellItem(Referenceable):
                 position=data.get(str, "Well Position", NOT_APPLICABLE),
                 well_location_identifier=data.get(str, "Well Position"),
                 quencher_dye_setting=data.get(str, "Quencher"),
-                sample_color=data.get(str, "Sample Color"),
-                biogroup_name=data.get(str, "Biogroup Name"),
-                biogroup_color=data.get(str, "Biogroup Color"),
-                target_color=data.get(str, "Target Color"),
+                group_identifier=data.get(str, "Biogroup Name"),
                 sample_role_type=data.get(str, "Task"),
+                extra_data={
+                    "well identifier": identifier,
+                    "sample color": data.get(str, "Sample Color"),
+                    "biogroup color": data.get(str, "Biogroup Color"),
+                    "target color": data.get(str, "Target Color"),
+                },
             ),
             WellItem(
                 uuid=random_uuid_str(),
@@ -170,11 +170,14 @@ class WellItem(Referenceable):
                 position=data.get(str, "Well Position", NOT_APPLICABLE),
                 well_location_identifier=data.get(str, "Well Position"),
                 quencher_dye_setting=data.get(str, "Quencher"),
-                sample_color=data.get(str, "Sample Color"),
-                biogroup_name=data.get(str, "Biogroup Name"),
-                biogroup_color=data.get(str, "Biogroup Color"),
-                target_color=data.get(str, "Target Color"),
+                group_identifier=data.get(str, "Biogroup Name"),
                 sample_role_type=data.get(str, "Task"),
+                extra_data={
+                    "well identifier": identifier,
+                    "sample color": data.get(str, "Sample Color"),
+                    "biogroup color": data.get(str, "Biogroup Color"),
+                    "target color": data.get(str, "Target Color"),
+                },
             ),
         )
 
@@ -194,11 +197,14 @@ class WellItem(Referenceable):
             position=data.get(str, "Well Position", NOT_APPLICABLE),
             well_location_identifier=data.get(str, "Well Position"),
             quencher_dye_setting=data.get(str, "Quencher"),
-            sample_color=data.get(str, "Sample Color"),
-            biogroup_name=data.get(str, "Biogroup Name"),
-            biogroup_color=data.get(str, "Biogroup Color"),
-            target_color=data.get(str, "Target Color"),
+            group_identifier=data.get(str, "Biogroup Name"),
             sample_role_type=data.get(str, "Task"),
+            extra_data={
+                "well identifier": identifier,
+                "sample color": data.get(str, "Sample Color"),
+                "biogroup color": data.get(str, "Biogroup Color"),
+                "target color": data.get(str, "Target Color"),
+            },
         )
 
 
@@ -360,7 +366,6 @@ class Result:
     baseline_corrected_reporter_result: float | None
     genotyping_determination_result: str | None
     genotyping_determination_method_setting: float | None
-    omit: bool | None
     quantity: float | None
     quantity_mean: float | None
     quantity_sd: float | None
@@ -379,13 +384,9 @@ class Result:
     slope: float | None
     efficiency: float | None
     comments: str | None
-    highsd: str | None
-    noamp: str | None
-    expfail: str | None
-    tholdfail: str | None
-    prfdrop: str | None
     amp_score: float | None
     cq_conf: float | None
+    extra_data: dict[str, Any] | None = None
 
     @staticmethod
     def create(
@@ -481,7 +482,6 @@ class Result:
                 genotyping_determination_method_setting=data.get(
                     float, "Threshold Value"
                 ),
-                omit=data.get(bool, "Omit"),
                 quantity=data.get(float, "Quantity"),
                 quantity_mean=data.get(float, "Quantity Mean"),
                 quantity_sd=data.get(float, "Quantity SD"),
@@ -500,13 +500,16 @@ class Result:
                 slope=data.get(float, "Slope"),
                 efficiency=data.get(float, "Efficiency"),
                 comments=data.get(str, "Comments"),
-                highsd=data.get(str, "HIGHSD"),
-                noamp=data.get(str, "NOAMP"),
-                expfail=data.get(str, "EXPFAIL"),
-                tholdfail=data.get(str, "THOLDFAIL"),
-                prfdrop=data.get(str, "PRFDROP"),
                 amp_score=data.get(float, "Amp Score"),
                 cq_conf=data.get(float, "Cq Conf"),
+                extra_data={
+                    "omit": data.get(bool, "Omit"),
+                    "highsd": data.get(str, "HIGHSD"),
+                    "noamp": data.get(str, "NOAMP"),
+                    "expfail": data.get(str, "EXPFAIL"),
+                    "tholdfail": data.get(str, "THOLDFAIL"),
+                    "prfdrop": data.get(str, "PRFDROP"),
+                },
             )
             for allele_prefix in allele_prefixes
         }