+++ CHANGELOG.md
@@ -602,8 +602,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor Agilent Gen5 with explicit dataclasses structure
-- Update Benchman Vi-cell Blu adapter to use the new cell-counting BENCHLING/2023/11 schema
-- Update Benchman Vi-cell XR adapter to use the new cell-counting BENCHLING/2023/11 schema
+- Update Beckman Vi-cell Blu adapter to use the new cell-counting BENCHLING/2023/11 schema
+- Update Beckman Vi-cell XR adapter to use the new cell-counting BENCHLING/2023/11 schema
 - Set mypy's disallow_any_generics to True. Ideally, new files should not suppress these warnings
 - Refactor way to extract and validate information from pandas series in AppBio QuantStudio
 - Simplify CSV lines reader